eklavya_backend/pyproject.toml
[project]
name = "eklavya-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "uvicorn[standard]>=0.41.0",
]
